mutualise prent validation function
