+++ api/src/controllers/young/inscription2023.js
@@ -7,7 +7,7 @@ const crypto = require("crypto");
 const YoungObject = require("../../models/young");
 const { capture } = require("../../sentry");
 const { serializeYoung } = require("../../utils/serializer");
-const { validateFirstName } = require("../../utils/validator");
+const { validateFirstName, validateParents } = require("../../utils/validator");
 const { ERRORS, STEPS2023, YOUNG_SITUATIONS } = require("../../utils");
 const {
   canUpdateYoungStatus,
@@ -15,7 +15,6 @@ const {
   YOUNG_STATUS,
   SENDINBLUE_TEMPLATES,
   isInRuralArea,
-  PHONE_ZONES_NAMES,
   PHONE_ZONES_NAMES_ARR,
   formatPhoneNumberFromPhoneZone,
 } = require("snu-lib");
@@ -91,14 +90,14 @@ router.put("/eligibilite", passport.authenticate("young", { session: false, fail
       ...value,
       ...(value.livesInFrance === "true"
         ? {
-            foreignCountry: "",
-            foreignAddress: "",
-            foreignCity: "",
-            foreignZip: "",
-            hostFirstName: "",
-            hostLastName: "",
-            hostRelationship: "",
-          }
+          foreignCountry: "",
+          foreignAddress: "",
+          foreignCity: "",
+          foreignZip: "",
+          hostFirstName: "",
+          hostLastName: "",
+          hostRelationship: "",
+        }
         : {}),
       ...validateCorrectionRequest(young, keyList),
     };
@@ -321,56 +320,7 @@ router.put("/representants/:type", passport.authenticate("young", { session: fal
     const { error: typeError, value: type } = checkParameter(req.params.type);
     if (typeError) return res.status(400).send({ ok: false, code: ERRORS.INVALID_PARAMS });
 
-    const isRequired = type !== "save";
-
-    const representantSchema = {
-      parent1Status: needRequired(Joi.string().trim().valid("father", "mother", "representant"), isRequired),
-      parent1FirstName: needRequired(validateFirstName().trim(), isRequired),
-      parent1LastName: needRequired(Joi.string().trim(), isRequired),
-      parent1Email: needRequired(Joi.string().trim().email(), isRequired),
-      parent1Phone: needRequired(Joi.string().trim(), isRequired),
-      parent1PhoneZone: needRequired(
-        Joi.string()
-          .trim()
-          .valid(...PHONE_ZONES_NAMES_ARR),
-        isRequired,
-      ),
-      parent2: needRequired(Joi.string().trim().valid(true, false), isRequired),
-      parent2Status: Joi.alternatives().conditional("parent2", {
-        is: true,
-        then: needRequired(Joi.string().trim().valid("father", "mother", "representant"), isRequired),
-        otherwise: Joi.isError(new Error()),
-      }),
-      parent2FirstName: Joi.alternatives().conditional("parent2", { is: true, then: needRequired(validateFirstName().trim(), isRequired), otherwise: Joi.isError(new Error()) }),
-      parent2LastName: Joi.alternatives().conditional("parent2", {
-        is: true,
-        then: needRequired(Joi.string().uppercase().trim(), isRequired),
-        otherwise: Joi.isError(new Error()),
-      }),
-      parent2Email: Joi.alternatives().conditional("parent2", {
-        is: true,
-        then: needRequired(Joi.string().trim().email(), isRequired),
-        otherwise: Joi.isError(new Error()),
-      }),
-      parent2Phone: Joi.alternatives().conditional("parent2", {
-        is: true,
-        then: needRequired(Joi.string().trim(), isRequired),
-        otherwise: Joi.isError(new Error()),
-      }),
-
-      parent2PhoneZone: Joi.alternatives().conditional("parent2", {
-        is: true,
-        then: needRequired(
-          Joi.string()
-            .trim()
-            .valid(...PHONE_ZONES_NAMES_ARR),
-          isRequired,
-        ),
-        otherwise: Joi.isError(new Error()),
-      }),
-    };
-
-    let { error, value } = Joi.object(representantSchema).validate(req.body, { stripUnknown: true });
+    let { error, value } = validateParents(req.body, type !== "save");
 
     if (error) {
       return res.status(400).send({ ok: false, code: ERRORS.INVALID_PARAMS });
